# Handle escaped keys and values in JSON