--- a/parser.go
+++ b/parser.go
@@ -9,13 +9,14 @@ import (
 
 // Errors
 var (
-	KeyPathNotFoundError  = errors.New("Key path not found")
-	UnknownValueTypeError = errors.New("Unknown value type")
-	MalformedJsonError    = errors.New("Malformed JSON error")
-	MalformedStringError  = errors.New("Value is string, but can't find closing '\"' symbol")
-	MalformedArrayError   = errors.New("Value is array, but can't find closing ']' symbol")
-	MalformedObjectError  = errors.New("Value looks like object, but can't find closing '}' symbol")
-	MalformedValueError   = errors.New("Value looks like Number/Boolean/None, but can't find its end: ',' or '}' symbol")
+	KeyPathNotFoundError       = errors.New("Key path not found")
+	UnknownValueTypeError      = errors.New("Unknown value type")
+	MalformedJsonError         = errors.New("Malformed JSON error")
+	MalformedStringError       = errors.New("Value is string, but can't find closing '\"' symbol")
+	MalformedArrayError        = errors.New("Value is array, but can't find closing ']' symbol")
+	MalformedObjectError       = errors.New("Value looks like object, but can't find closing '}' symbol")
+	MalformedValueError        = errors.New("Value looks like Number/Boolean/None, but can't find its end: ',' or '}' symbol")
+	MalformedStringEscapeError = errors.New("Encountered an invalid escape sequence in a string")
 )
 
 func tokenEnd(data []byte) int {