Added proper JSON escaped string handling in both keys and values (as per [RFC 7159](https://tools.ietf.org/html/rfc7159)). These changes necessarily resulted in some performance degradation, but it's relatively minor (< 10%, it seems).

(Note: this is a refinement of #39)

*Benchmark before:*
```
BenchmarkJsonParserLarge-8 	 	     64506 ns/op	       0 allocs/op
BenchmarkJsonParserMedium-8	 	     10865 ns/op	       0 allocs/op
BenchmarkJsonParserSmall-8 		      1026 ns/op	       0 allocs/op
```

*Benchmark after:*
```
BenchmarkJsonParserLarge-8 	  	     66558 ns/op	       0 allocs/op
BenchmarkJsonParserMedium-8	 	     11958 ns/op	       0 allocs/op
BenchmarkJsonParserSmall-8 		      1061 ns/op	       0 allocs/op
```