@daboyuka 2016-04-11
Hmm, performance is pretty bad after this change. Apparently `bytes.IndexByte` is not as fast as I'd hoped. Will see if I can speed this up some before merge.