@buger 2016-04-13
Also i guess it should use https://github.com/buger/jsonparser/tree/unescaped-keys for escape detection.