@daboyuka 2016-04-13
Right, that is what I need to change to improve performance.

My original method was to use `bytes.Index(key, '\\')` to determine if any backslashes exist in the key, but this is apparently far slower than adding checking to `stringEnd` like you did. I would have thought the `bytes` method would be fast due to optimized CPU instructions, but apparently not :smile:.

Will try to finish this up soon.