@buger 2016-04-13
I had same issue about bytes.Index and find out that it definitly very fast, but only when you need to call it once, using just as true/false to detect if string include byte. If you call bytes.Index multiple times, its overhead is bigger then performance gain.