@daboyuka 2016-04-13
Yeah, I think it's optimized for large chunks of bytes, so it's generally unsuitable for, i.e., JSON keys.