@daboyuka 2016-04-13
I'm rebasing this onto master with the key-unescape branch change (checking for escaping in `stringEnd`).